+++ src/pytest_rich/header.py
@@ -1,4 +1,5 @@
 import sys
+from typing import Iterable
 from typing import Union
 
 import pytest
@@ -14,6 +15,7 @@ def generate_header_panel(session: Session) -> Panel:
         _generate_sysinfo_col(),
         _generate_root_col(session),
         _generate_plugins_col(session),
+        *_generate_header_hook_cols(session),
     ]
 
     return Panel(Group(*columns))
@@ -50,3 +52,16 @@ def _generate_plugins_col(session: Session) -> Union[Columns, None]:
             f"plugins [cyan]{', '.join(_plugin_nameversions(plugins))}",
         ]
     )
+
+
+def _generate_header_hook_cols(session: Session) -> Iterable[Columns]:
+    lines = session.config.hook.pytest_report_header(
+        config=session.config, start_path=session.config.invocation_params.dir
+    )
+
+    for line_or_lines in reversed(lines):
+        if isinstance(line_or_lines, str):
+            yield Columns([line_or_lines])
+        else:
+            for line in line_or_lines:
+                yield Columns([line])